Fetch site files from the malariaverse Packit server.
This adds `fetch_site` and `fetch_files` function to pull data from the
malariaverse Packit instance. It uses a cache in the user's home
directory to save files and avoid unnecessary repeated downloads.

The user needs to be authenticated in order to access the Packit server.
This is done interactively the first time a user tries to call these
functions. The user is prompted to open GitHub in their browser and type
in a short code, after which their orderly client will be able to
connect to Packit. The token that is obtained in cached to remove the
need to re-authenticate. Alternatively, users may set a GITHUB_TOKEN
environment variable with a suitable GitHub personal access token.